+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+* add **SparqlSourceProcessor**: Extracts an RDF document from a SPARQL endpoint.
+
 ## [v0.2.1] - 2020-12-03
 
 ### Fixed